chore(constants): move logger constants to their impl
