--- src/constants/index.js
@@ -1,14 +1,0 @@
-// Add all labels here
-const LOGGER_LABELS = {
-  DEFAULT: 'DEFAULT',
-  ROUTER: 'ROUTER',
-  CONTROLLER: 'CONTROLLER',
-  SERVICE: 'SERVICE',
-  ADD_USE_CASE: 'ADD_USE_CASE',
-  GET_ALL_USE_CASE: 'GET_ALL_USE_CASE',
-  CHANGE_NAME_USE_CASE: 'CHANGE_NAME_USE_CASE',
-  DATA_SOURCE: 'DATA_SOURCE',
-}
-
-
-module.exports = { LOGGER_LABELS }
